Abstract out the api logic from actions
