+++ src/actions/teams.js
@@ -1,8 +1,7 @@
 // Copyright (c) 2016 Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import _ from 'lodash';
-import {requestData, requestSuccess, requestFailure} from './helpers.js';
+import {bindClientFunc} from './helpers.js';
 import Client from 'client/client_instance';
 import {TeamsTypes as types} from 'constants';
 
@@ -15,35 +14,10 @@ export function selectTeam(team) {
 }
 
 export function fetchTeams() {
-    return async (dispatch) => {
-        try {
-            dispatch(requestData(types.FETCH_TEAMS_REQUEST));
-            const url = `${Client.getTeamsRoute()}/all_team_listings`;
-            const resp = await fetch(url, {
-                headers: {
-                    Authorization: 'Bearer bpfzjdtxybddtnfptmt78hghqo'
-                }
-            });
-            let data;
-            const contentType = _.first(resp.headers.map['content-type']) || 'unknown';
-            if (contentType === 'application/json') {
-                data = await resp.json();
-            } else {
-                data = await resp.text();
-            }
-            if (resp.ok) {
-                dispatch(requestSuccess(types.FETCH_TEAMS_SUCCESS, data));
-            } else {
-                let msg;
-                if (contentType === 'application/json') {
-                    msg = data.message;
-                } else {
-                    msg = data;
-                }
-                throw new Error(msg);
-            }
-        } catch (err) {
-            dispatch(requestFailure(types.FETCH_TEAMS_FAILURE, err));
-        }
-    };
+    return bindClientFunc(
+        Client.fetchTeams,
+        types.FETCH_TEAMS_REQUEST,
+        types.FETCH_TEAMS_SUCCESS,
+        types.FETCH_TEAMS_FAILURE
+    );
 }
