+++ src/client/client.js
@@ -1,6 +1,8 @@
 // Copyright (c) 2016 Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
+import _ from 'lodash';
+
 const HEADER_AUTH = 'Authorization';
 const HEADER_BEARER = 'BEARER';
 const HEADER_REQUESTED_WITH = 'X-Requested-With';
@@ -238,6 +240,17 @@ export default class Client {
         );
     }
 
+    fetchChannels = (onRequest, onSuccess, onFailure) => {
+        return this.doFetch(
+            `${this.getChannelsRoute()}/`,
+            {method: 'get'},
+            onRequest,
+            onSuccess,
+            onFailure
+        );
+    }
+
+
     // Post routes
 
     createPost = (post, onRequest, onSuccess, onFailure) => {
@@ -255,13 +268,24 @@ export default class Client {
             onRequest();
         }
         try {
-            const response = await fetch(url, this.getOptions(options));
-            const data = await response.json();
-            if (response.ok) {
-                return onSuccess(data, response);
+            const resp = await fetch(url, this.getOptions(options));
+            let data;
+            const contentType = _.first(resp.headers.map['content-type']) || 'unknown';
+            if (contentType === 'application/json') {
+                data = await resp.json();
+            } else {
+                data = await resp.text();
             }
-            const {message} = data;
-            throw new Error(message || 'Failed to fetch');
+            if (resp.ok) {
+                return onSuccess(data, resp);
+            }
+            let msg;
+            if (contentType === 'application/json') {
+                msg = data.message;
+            } else {
+                msg = data;
+            }
+            throw new Error(msg);
         } catch (err) {
             if (this.logToConsole) {
                 console.log(err); // eslint-disable-line no-console
